# console/src/api/generated.ts
// Generated by scripts/unified_spec.py. Do not edit.
export type Apply = { generation: number; session_id: string; subject_digest: string };
export type Capture = { generation: number; source_turn_id?: string | null; origin?: "automatic" | "explicit" | "salvage" };
export type ConnectionCreate = { kind: "modal" | "github" | "opencode_zen"; label?: string; credential: Record<string, string> };
export type ConnectionView = { id: string; workspace_id: string; kind: string; label: string; state: string; version: number; health: string };
export type Continuation = { summary: string; title?: string; changeset_id?: string | null };
export type CredentialReplace = { expected_version: number; credential: Record<string, string> };
export type DeliveryCreate = { transport?: "pull_request" | "git_branch" | "export"; github_connection_id?: string | null; base_ref?: string; automatic?: boolean };
export type EmailRequest = { email: string };
export type FileWrite = { generation: number; path: string; content: string; digest?: string | null };
export type Generation = { generation: number };
export type HTTPValidationError = { detail?: (ValidationError)[] };
export type Login = { email: string; password: string };
export type Merge = { expected_version: number; subject_digest: string; expected_head: string; expected_base?: string | null; method?: "squash" | "merge" | "rebase" };
export type Message = { content: string; routing?: "queue" | "note" | "steer"; settings?: Record<string, unknown> };
export type MessageView = { id: string; content: string; ordinal: number; routing: string };
export type PartView = { id: string; execution_id: string; kind: string; revision: number; content: string };
export type PasswordChange = { current_password: string; password: string };
export type PasswordReset = { verifier: string; password: string };
export type ProjectCreate = { name?: string; repository: string; base_ref?: string; spec?: Record<string, unknown> };
export type ProjectPublish = { name?: string; repository: string; base_ref?: string; spec?: Record<string, unknown>; expected_version: number };
export type Publish = { turn_id: string };
export type SessionCreate = { title?: string; project_version_id?: string | null; repository?: string | null; base_ref?: string; provider_id?: string; backend?: "modal" | "local"; model?: string; modal_connection_id?: string | null; github_connection_id?: string | null; zen_connection_id?: string | null; message?: Message | null };
export type SessionPatch = { expected_version: number; title: string };
export type SessionView = { id: string; workspace_id: string; creator_id: string; title: string; role: string; lifecycle: string; version: number; provider_id: string; project_version_id?: string | null; effective_inputs: Record<string, unknown>; event_watermark: number; worktree: WorktreeView; turns: (TurnView)[]; messages: (MessageView)[]; parts: (PartView)[] };
export type Spawn = { budget_seconds?: number; changeset_id: string; role?: "review" | "test" | "research" | "integration"; summary?: string; model?: string | null };
export type TerminalOpen = { command?: (string)[] };
export type TurnView = { id: string; ordinal: number; session_id: string; state: string; evidence_complete: boolean; settings: Record<string, unknown>; outcome?: Record<string, unknown> | null; reason?: string | null };
export type ValidationError = { loc: (string | number)[]; msg: string; type: string; input?: unknown; ctx?: Record<string, unknown> };
export type Version = { expected_version: number };
export type Wait = { seconds?: number };
export type WorktreeView = { id: string; generation: number; availability: string; last_snapshot_id?: string | null; base_sha?: string | null };