# docs/specs/unified/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  title: SBX unified business API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/operations/{oid}:
    get:
      operationId: operation_api_operations__oid__get
      parameters:
      - in: path
        name: oid
        required: true
        schema:
          title: Oid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Operation
...